Schedule y logs

+++ src/main/java/coop/magnesium/sulfur/system/StartupBean.java
@@ -14,13 +14,10 @@ import javax.ejb.*;
 import javax.enterprise.event.Event;
 import javax.inject.Inject;
 import java.time.*;
-import java.time.temporal.ChronoUnit;
 import java.util.Date;
 import java.util.List;
 import java.util.UUID;
 import java.util.logging.Logger;
-
-import static coop.magnesium.sulfur.system.TimerType.*;
 
 /**
  * Created by rsperoni on 18/11/17.
@@ -65,35 +62,7 @@ public class StartupBean {
         }
         configuraciones();
         setMyselfAsNodoMaster();
-        setTimerNotificaciones();
-        setTimerCleanRecuperacionContrasena();
-        setTimerEnvioMails();
 
-    }
-
-    public void setTimerNotificaciones() {
-        if (configuracionDao.getPeriodicidadNotificaciones() != 0) {
-            Instant instant = Instant.now().plus(configuracionDao.getPeriodicidadNotificaciones(), ChronoUnit.HOURS);
-            TimerConfig timerConfig = new TimerConfig();
-            timerConfig.setInfo(new DataTimer(TimerType.NOTIFICACION_ALERTA, null));
-            timerService.createSingleActionTimer(Date.from(instant), timerConfig);
-        }
-    }
-
-    public void setTimerCleanRecuperacionContrasena() {
-        Instant instant = Instant.now().plus(1, ChronoUnit.HOURS);
-        TimerConfig timerConfig = new TimerConfig();
-        timerConfig.setInfo(new DataTimer(TimerType.RECUPERACION_CONTRASENA, null));
-        timerService.createSingleActionTimer(Date.from(instant), timerConfig);
-    }
-
-    public void setTimerEnvioMails() {
-        TimerConfig timerConfig = new TimerConfig();
-        timerConfig.setInfo(new DataTimer(TimerType.ENVIO_MAIL, null));
-        Instant instant = LocalDateTime.of(LocalDate.now().plusDays(1), LocalTime.of(9, 30)).atZone(ZoneId.systemDefault()).toInstant();
-        Date dateFromOld = Date.from(instant);
-        logger.info("Set TIMER ENVIO MAILS: " + dateFromOld.toString());
-        timerService.createSingleActionTimer(dateFromOld, timerConfig);
     }
 
     public void setMyselfAsNodoMaster() {
@@ -122,33 +91,6 @@ public class StartupBean {
         }
     }
 
-    @Timeout
-    public void timeout(Timer timer) {
-        //Solo si soy master
-        if (configuracionDao.getNodoMaster().equals(jbossNodeName)) {
-            if (timer.getInfo().getClass().getCanonicalName().equals(DataTimer.class.getCanonicalName())) {
-                DataTimer dataTimer = ((DataTimer) timer.getInfo());
-                switch (dataTimer.timerType) {
-                    case NOTIFICACION_ALERTA:
-                        logger.info("Timeout: " + NOTIFICACION_ALERTA.name());
-                        alertaHorasSinCargar();
-                        setTimerNotificaciones();
-                        break;
-                    case RECUPERACION_CONTRASENA:
-                        logger.info("Timeout: " + RECUPERACION_CONTRASENA.name());
-                        cleanRecuperacionContrasena();
-                        setTimerCleanRecuperacionContrasena();
-                        break;
-                    case ENVIO_MAIL:
-                        logger.info("Timeout: " + ENVIO_MAIL.name());
-                        enviarMailsConNotificaciones();
-                        setTimerEnvioMails();
-                        break;
-                }
-            }
-        }
-    }
-
     public void putRecuperacionPassword(RecuperacionPassword recuperacionPassword) {
         Instant instant = recuperacionPassword.getExpirationDate().toInstant(ZoneOffset.UTC);
         TimerConfig timerConfig = new TimerConfig();
@@ -167,49 +109,59 @@ public class StartupBean {
         }
     }
 
+    @Schedule(hour = "*", info = "cleanRecuperacionContrasena", persistent = false)
     public void cleanRecuperacionContrasena() {
         //Solo si soy master
-
-        logger.info("Master cleaning Recuperacion Contraseña");
-        recuperacionPasswordDao.findAll().forEach(recuperacionPassword -> {
-            if (recuperacionPassword.getExpirationDate().isBefore(LocalDateTime.now())) {
-                recuperacionPasswordDao.delete(recuperacionPassword);
-            }
-        });
-
+        if (configuracionDao.getNodoMaster().equals(jbossNodeName)) {
+            logger.info("Master cleaning Recuperacion Contraseña");
+            recuperacionPasswordDao.findAll().forEach(recuperacionPassword -> {
+                if (recuperacionPassword.getExpirationDate().isBefore(LocalDateTime.now())) {
+                    recuperacionPasswordDao.delete(recuperacionPassword);
+                }
+            });
+        }
     }
 
+    @Schedule(hour = "*/24", minute = "*", second = "*/5", info = "alertaHorasSinCargar", persistent = false)
     public void alertaHorasSinCargar() {
-
-        logger.info("Master generando notificaciones");
-        LocalDate hoy = LocalDate.now();
-        //Solo días de semana
-        if (!hoy.getDayOfWeek().equals(DayOfWeek.SATURDAY) && !hoy.getDayOfWeek().equals(DayOfWeek.SUNDAY)) {
-            colaboradorDao.findAll().stream().filter(colaborador -> colaborador.getCargo() != null)
-                    .forEach(colaborador -> {
-                        //Si hace más de 3 días que no hay horas
-                        if (horaDao.findAllByColaborador(colaborador, LocalDate.now().minusDays(3), LocalDate.now()).isEmpty()) {
-                            notificacionEvent.fire(new Notificacion(TipoNotificacion.FALTAN_HORAS, colaborador, colaborador.getNombre() + " no cargó horas en más de 2 días."));
-                        }
-                    });
+        //Solo si soy master
+        if (configuracionDao.getNodoMaster().equals(jbossNodeName)) {
+            logger.info("Master generando notificaciones");
+            LocalDate hoy = LocalDate.now();
+            //Solo días de semana
+            if (!hoy.getDayOfWeek().equals(DayOfWeek.SATURDAY) && !hoy.getDayOfWeek().equals(DayOfWeek.SUNDAY)) {
+                colaboradorDao.findAll().stream().filter(colaborador -> colaborador.getCargo() != null)
+                        .forEach(colaborador -> {
+                            //Si hace más de 3 días que no hay horas
+                            if (horaDao.findAllByColaborador(colaborador, LocalDate.now().minusDays(3), LocalDate.now()).isEmpty()) {
+                                notificacionEvent.fire(new Notificacion(TipoNotificacion.FALTAN_HORAS, colaborador, colaborador.getNombre() + " no cargó horas en más de 2 días."));
+                            }
+                        });
+            }
         }
 
     }
 
+    @Schedule(dayOfWeek = "Mon,Wed,Fri", hour = "9", minute = "30", info = "enviarMailsConNotificaciones", persistent = false)
     public void enviarMailsConNotificaciones() {
-        logger.info("Master enviando mails");
-        StringBuilder stringBuilder = new StringBuilder();
-        notificacionDao.findAllNoEnviadas().forEach(notificacion -> {
-            stringBuilder.append("- ").append(notificacion.getTexto()).append("\n");
-            notificacion.setEnviado(true);
-        });
+        //Solo si soy master
+        if (configuracionDao.getNodoMaster().equals(jbossNodeName)) {
+            logger.info("Master enviando mails");
+            StringBuilder stringBuilder = new StringBuilder();
+            notificacionDao.findAllNoEnviadas().stream()
+                    .filter(notificacion -> notificacion.getTipo().equals(TipoNotificacion.FALTAN_HORAS))
+                    .forEach(notificacion -> {
+                stringBuilder.append("- ").append(notificacion.getTexto()).append("\n");
+                notificacion.setEnviado(true);
+            });
 
-        List<String> mailsAdmins = configuracionDao.getDestinatariosNotificacionesAdmins();
-        if (!stringBuilder.toString().isEmpty()) {
-            mailEvent.fire(
-                    new MailEvent(mailsAdmins,
-                            stringBuilder.toString(),
-                            "MARQ: Notificaciones"));
+            List<String> mailsAdmins = configuracionDao.getDestinatariosNotificacionesAdmins();
+            if (!stringBuilder.toString().isEmpty()) {
+                mailEvent.fire(
+                        new MailEvent(mailsAdmins,
+                                stringBuilder.toString(),
+                                "MARQ: Notificaciones"));
+            }
         }
 
     }
